Hide the boot splash instead of detaching it: removal crashed every navigation
BootSplashRemover called .remove() on the boot splash -- a node React
server-rendered as part of the layout's body and still owns. Detaching a
React-owned node outside React corrupts the reconciler's view of <body>,
so the next client-side navigation died in insertBefore/removeChild
("not a child of this node") and every page rendered the error boundary.

BootSplashHider sets an inline display:none instead (which wins over the
stylesheet's display:flex) and changes no tree structure. The regression
test asserts the splash is still attached exactly where the server put it
-- the original test asserted the opposite, which is what let the defect
through. Verified in a real browser: repeated login/register round-trip
navigations with zero page errors.
